+++ CHANGELOG.md
@@ -15,4 +15,4 @@
  * fix file delete
  * fix sqlFileUtils包问题
  * fix create temp file
- 
+ * fix file extension